requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.12.0
asgiref==3.11.0
asyncpg==0.31.0
attrs==25.4.0
beautifulsoup4==4.14.3
black==25.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
db==0.1.1
Django==6.0
django-scheduler==0.10.1
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
icalendar==6.3.2
idna==3.11
isort==7.0.0
lxml==6.0.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
platformdirs==4.5.1
playwright==1.57.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
requests==2.32.5
selenium==4.39.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.45
sqlparse==0.5.4
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
yarl==1.22.0
